commit lavoro fino a 30-11-2019

+++ Models/Entities/Passwords.cs
@@ -5,11 +5,11 @@ namespace PasswordManager.Models.Entities
 {
     public partial class Passwords
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Password { get; set; }
         public string Descrizione { get; set; }
         public string DataInserimento { get; set; }
-        public long FkUtente { get; set; }
+        public int FkUtente { get; set; }
         public string Sito { get; set; }
         public string Tipo { get; set; }
     }
